Rubydoctest used to test hello method
Covers naming arguments,testing incrementally
and testing development.

Minimal doctest syntax, we use this because it is basically an IRB
session, but it records our knowledge.  If something in the language
causes our tests to break, then our knowledge is no longer 'good', we
can be alerted that something with Ruby has changed.

Using tests, we can ask Ruby by doing, by writing the code we think
works, and it will let us know that it is working or not.

Let the test and let Ruby talk to us.