Lookup Table exercise for barewords
The local variable and method names and keywords all look the same,
though there may be uses that make it unambiguous. They are all 'bare
words'.

For example, a variable can not take an argument, and so if Ruby sees a
bareword that takes an argument, it will not look in the "variable"
column of the lookup table, but will know that it is a method you are
trying to invoke.