Working with iterators  on Strings
We got more familiar with the return of each object that is returned
when we call a method, and realize that we start working with a string,
end up working with an Enumerator, and finally an array, then back to
String for our return.

Went over use of RI (Ruby Interactive) for documentation at the
keyboard.

Also quick experiments in IRB to flesh out some ideas.